Add "random communication benchmark" tool
git-svn-id: https://svn.apache.org/repos/asf/incubator/hama/trunk@1072853 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hama/bsp/BSPJob.java
@@ -140,7 +140,7 @@ public class BSPJob extends BSPJobContext {
   // /////////////////////////////////////
   // Methods for Job Control
   // /////////////////////////////////////
-  public float progress() throws IOException {
+  public long progress() throws IOException {
     ensureState(JobState.RUNNING);
     return info.progress();
   }
